:sparkles:  feat: refactor gcd
