+++ packages/gcd/README.md
@@ -1,6 +1,6 @@
 <header>
   <h1 align="center">
-    <a href="https://github.com/guanghechen/algorithm.ts/tree/release-2.x.x/packages/gcd#readme">@algorithm.ts/gcd</a>
+    <a href="https://github.com/guanghechen/algorithm.ts/tree/release-3.x.x/packages/gcd#readme">@algorithm.ts/gcd</a>
   </h1>
   <div align="center">
     <a href="https://www.npmjs.com/package/@algorithm.ts/gcd">
@@ -72,14 +72,8 @@ A typescript implementation of the **gcd** and **Extended Euclidean** algorithm.
   yarn add @algorithm.ts/gcd
   ```
 
-* deno
-
-  ```typescript
-  import { gcd, euclidean } from 'https://raw.githubusercontent.com/guanghechen/algorithm.ts/main/packages/gcd/src/index.ts'
-  ```
 
 ## Usage
-
 
 * gcd
 
@@ -107,5 +101,5 @@ A typescript implementation of the **gcd** and **Extended Euclidean** algorithm.
 * [数论基础之模方程初步][gcd]
 
 
-[homepage]: https://github.com/guanghechen/algorithm.ts/tree/release-2.x.x/packages/gcd#readme
+[homepage]: https://github.com/guanghechen/algorithm.ts/tree/release-3.x.x/packages/gcd#readme
 [gcd]: https://me.guanghechen.com/post/math/number-theory/%E6%A8%A1%E6%96%B9%E7%A8%8B/basic/
